pyproject.toml
[project]
name = "eusougraciosa-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.14",
    "passlib>=1.7.4",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
